+++ src/lib/helpers.test.ts
@@ -1,15 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
-import { Binding, CapabilityExport } from "./types";
-import { Event } from "./enums";
+import { CapabilityExport } from "./types";
 import {
   addVerbIfNotExists,
   bindingAndCapabilityNSConflict,
   createDirectoryIfNotExists,
   createRBACMap,
   checkDeploymentStatus,
-  filterNoMatchReason,
   dedent,
   generateWatchNamespaceError,
   hasAnyOverlap,
@@ -30,10 +28,8 @@ import * as fc from "fast-check";
 import { expect, describe, test, jest, beforeEach, afterEach } from "@jest/globals";
 import { promises as fs } from "fs";
 import { SpiedFunction } from "jest-mock";
-import { K8s, GenericClass, KubernetesObject, kind } from "kubernetes-fluent-client";
+import { K8s, GenericClass, KubernetesObject } from "kubernetes-fluent-client";
 import { K8sInit } from "kubernetes-fluent-client/dist/fluent/types";
-
-export const callback = () => undefined;
 
 jest.mock("kubernetes-fluent-client", () => {
   return {
@@ -1065,331 +1061,6 @@ describe("replaceString", () => {
     const stringB = "";
     const expected = "Remove.";
     expect(replaceString(original, stringA, stringB)).toBe(expected);
-  });
-});
-
-describe("filterNoMatchReason", () => {
-  test("returns regex namespace filter error for Pods whos namespace does not match the regex", () => {
-    const binding = {
-      kind: { kind: "Pod" },
-      filters: { regexNamespaces: ["(.*)-system"], namespaces: [] },
-    };
-    const obj = { metadata: { namespace: "pepr-demo" } };
-    const objArray = [
-      { ...obj },
-      { ...obj, metadata: { namespace: "pepr-uds" } },
-      { ...obj, metadata: { namespace: "pepr-core" } },
-      { ...obj, metadata: { namespace: "uds-ns" } },
-      { ...obj, metadata: { namespace: "uds" } },
-    ];
-    const capabilityNamespaces: string[] = [];
-    objArray.map(object => {
-      const result = filterNoMatchReason(
-        binding as unknown as Partial<Binding>,
-        object as unknown as Partial<KubernetesObject>,
-        capabilityNamespaces,
-      );
-      expect(result).toEqual(
-        `Ignoring Watch Callback: Binding defines namespace regexes '["(.*)-system"]' but Object carries '${object?.metadata?.namespace}'.`,
-      );
-    });
-  });
-
-  test("returns no regex namespace filter error for Pods whos namespace does match the regex", () => {
-    const binding = {
-      kind: { kind: "Pod" },
-      filters: { regexNamespaces: [/(.*)-system/], namespaces: [] },
-    };
-    const obj = { metadata: { namespace: "pepr-demo" } };
-    const objArray = [
-      { ...obj, metadata: { namespace: "pepr-system" } },
-      { ...obj, metadata: { namespace: "pepr-uds-system" } },
-      { ...obj, metadata: { namespace: "uds-system" } },
-      { ...obj, metadata: { namespace: "some-thing-that-is-a-system" } },
-      { ...obj, metadata: { namespace: "your-system" } },
-    ];
-    const capabilityNamespaces: string[] = [];
-    objArray.map(object => {
-      const result = filterNoMatchReason(
-        binding as unknown as Partial<Binding>,
-        object as unknown as Partial<KubernetesObject>,
-        capabilityNamespaces,
-      );
-      expect(result).toEqual(``);
-    });
-  });
-
-  // Names Fail
-  test("returns regex name filter error for Pods whos name does not match the regex", () => {
-    const binding = {
-      kind: { kind: "Pod" },
-      filters: { regexName: "^system", namespaces: [] },
-    };
-    const obj = { metadata: { name: "pepr-demo" } };
-    const objArray = [
-      { ...obj },
-      { ...obj, metadata: { name: "pepr-uds" } },
-      { ...obj, metadata: { name: "pepr-core" } },
-      { ...obj, metadata: { name: "uds-ns" } },
-      { ...obj, metadata: { name: "uds" } },
-    ];
-    const capabilityNamespaces: string[] = [];
-    objArray.map(object => {
-      const result = filterNoMatchReason(
-        binding as unknown as Partial<Binding>,
-        object as unknown as Partial<KubernetesObject>,
-        capabilityNamespaces,
-      );
-      expect(result).toEqual(
-        `Ignoring Watch Callback: Binding defines name regex '^system' but Object carries '${object?.metadata?.name}'.`,
-      );
-    });
-  });
-
-  // Names Pass
-  test("returns no regex name filter error for Pods whos name does match the regex", () => {
-    const binding = {
-      kind: { kind: "Pod" },
-      filters: { regexName: /^system/, namespaces: [] },
-    };
-    const obj = { metadata: { name: "pepr-demo" } };
-    const objArray = [
-      { ...obj, metadata: { name: "systemd" } },
-      { ...obj, metadata: { name: "systemic" } },
-      { ...obj, metadata: { name: "system-of-kube-apiserver" } },
-      { ...obj, metadata: { name: "system" } },
-      { ...obj, metadata: { name: "system-uds" } },
-    ];
-    const capabilityNamespaces: string[] = [];
-    objArray.map(object => {
-      const result = filterNoMatchReason(
-        binding as unknown as Partial<Binding>,
-        object as unknown as Partial<KubernetesObject>,
-        capabilityNamespaces,
-      );
-      expect(result).toEqual(``);
-    });
-  });
-
-  test("returns namespace filter error for namespace objects with namespace filters", () => {
-    const binding = {
-      kind: { kind: "Namespace" },
-      filters: { namespaces: ["ns1"] },
-    };
-    const obj = {};
-    const capabilityNamespaces: string[] = [];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual("Ignoring Watch Callback: Cannot use namespace filter on a namespace object.");
-  });
-
-  test("return an Ignoring Watch Callback string if the binding name and object name are different", () => {
-    const binding = {
-      filters: { name: "pepr" },
-    };
-    const obj = {
-      metadata: {
-        name: "not-pepr",
-      },
-    };
-    const capabilityNamespaces: string[] = [];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual(`Ignoring Watch Callback: Binding defines name 'pepr' but Object carries 'not-pepr'.`);
-  });
-  test("returns no Ignoring Watch Callback string if the binding name and object name are the same", () => {
-    const binding = {
-      filters: { name: "pepr" },
-    };
-    const obj = {
-      metadata: { name: "pepr" },
-    };
-    const capabilityNamespaces: string[] = [];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual("");
-  });
-
-  test("return deletionTimestamp error when there is no deletionTimestamp in the object", () => {
-    const binding = {
-      filters: { deletionTimestamp: true },
-    };
-    const obj = {
-      metadata: {},
-    };
-    const capabilityNamespaces: string[] = [];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual("Ignoring Watch Callback: Binding defines deletionTimestamp but Object does not carry it.");
-  });
-
-  test("return no deletionTimestamp error when there is a deletionTimestamp in the object", () => {
-    const binding = {
-      filters: { deletionTimestamp: true },
-    };
-    const obj = {
-      metadata: {
-        deletionTimestamp: "2021-01-01T00:00:00Z",
-      },
-    };
-    const capabilityNamespaces: string[] = [];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).not.toEqual("Ignoring Watch Callback: Binding defines deletionTimestamp Object does not carry it.");
-  });
-
-  test("returns label overlap error when there is no overlap between binding and object labels", () => {
-    const binding = {
-      filters: { labels: { key: "value" } },
-    };
-    const obj = {
-      metadata: { labels: { anotherKey: "anotherValue" } },
-    };
-    const capabilityNamespaces: string[] = [];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual(
-      `Ignoring Watch Callback: Binding defines labels '{"key":"value"}' but Object carries '{"anotherKey":"anotherValue"}'.`,
-    );
-  });
-
-  test("returns annotation overlap error when there is no overlap between binding and object annotations", () => {
-    const binding = {
-      filters: { annotations: { key: "value" } },
-    };
-    const obj = {
-      metadata: { annotations: { anotherKey: "anotherValue" } },
-    };
-    const capabilityNamespaces: string[] = [];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual(
-      `Ignoring Watch Callback: Binding defines annotations '{"key":"value"}' but Object carries '{"anotherKey":"anotherValue"}'.`,
-    );
-  });
-
-  test("returns capability namespace error when object is not in capability namespaces", () => {
-    const binding = {
-      model: kind.Pod,
-      event: Event.Any,
-      kind: {
-        group: "",
-        version: "v1",
-        kind: "Pod",
-      },
-      filters: {
-        name: "bleh",
-        namespaces: [],
-        regexNamespaces: [],
-        regexName: "",
-        labels: {},
-        annotations: {},
-        deletionTimestamp: false,
-      },
-      callback,
-    };
-
-    const obj = {
-      metadata: { namespace: "ns2", name: "bleh" },
-    };
-    const capabilityNamespaces = ["ns1"];
-    const result = filterNoMatchReason(
-      binding as Binding,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual(
-      `Ignoring Watch Callback: Object carries namespace 'ns2' but namespaces allowed by Capability are '["ns1"]'.`,
-    );
-  });
-
-  test("returns binding namespace error when filter namespace is not part of capability namespaces", () => {
-    const binding = {
-      filters: { namespaces: ["ns3"], regexNamespaces: [] },
-    };
-    const obj = {};
-    const capabilityNamespaces = ["ns1", "ns2"];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual(
-      `Ignoring Watch Callback: Binding defines namespaces ["ns3"] but namespaces allowed by Capability are '["ns1","ns2"]'.`,
-    );
-  });
-
-  test("returns binding and object namespace error when they do not overlap", () => {
-    const binding = {
-      filters: { namespaces: ["ns1"], regexNamespaces: [] },
-    };
-    const obj = {
-      metadata: { namespace: "ns2" },
-    };
-    const capabilityNamespaces = ["ns1", "ns2"];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual(`Ignoring Watch Callback: Binding defines namespaces '["ns1"]' but Object carries 'ns2'.`);
-  });
-
-  test("return watch violation message when object is in an ignored namespace", () => {
-    const binding = {
-      filters: { namespaces: ["ns3"] },
-    };
-    const obj = {
-      metadata: { namespace: "ns3" },
-    };
-    const capabilityNamespaces = ["ns3"];
-    const ignoredNamespaces = ["ns3"];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-      ignoredNamespaces,
-    );
-    expect(result).toEqual(
-      `Ignoring Watch Callback: Object carries namespace 'ns3' but ignored namespaces include '["ns3"]'.`,
-    );
-  });
-
-  test("returns empty string when all checks pass", () => {
-    const binding = {
-      filters: { namespaces: ["ns1"], labels: { key: "value" }, annotations: { key: "value" } },
-    };
-    const obj = {
-      metadata: { namespace: "ns1", labels: { key: "value" }, annotations: { key: "value" } },
-    };
-    const capabilityNamespaces = ["ns1"];
-    const result = filterNoMatchReason(
-      binding as unknown as Partial<Binding>,
-      obj as unknown as Partial<KubernetesObject>,
-      capabilityNamespaces,
-    );
-    expect(result).toEqual("");
   });
 });
